Upgrade (v6.3.3 -> v7.0.0)
Ref: https://github.com/Chocobozzz/PeerTube/releases/tag/v7.0.0

We're not doing anything with this instruction:

> Ensure you have storage.original_video_files set in your configuration file

We configure PeerTube via environment variables and assume there's a
sensible default value for this configuration setting.
It only seems to be relevant if some "Keep a version of the input file"
setting is explicitly enabled as well, so most users should not be
affected anyway.